+++ src/main/kotlin/com/rodwan/coreapi/dtos/CartItemsResponse.kt
@@ -1,12 +1,13 @@
 package com.rodwan.coreapi.dtos
 
+import java.math.BigDecimal
 import java.time.Instant
 
 data class CartItemsResponse(
     val cartId: String,
     val userId: String,
     val items: List<CartItemDto>,
-    val totalAmount: Double,
+    val totalAmount: BigDecimal,
     val itemCount: Int,
     val lastModified: Instant,
 )
